# WIP: Experiment. Optimize 3. Use matrix-vector multiplications in Ax,Ay,Az directions