0. (Rebase on ObbHasOverlapBenchMark #14621)
1. (Branch out of ObbHasOverlapAVXTransformOnly) (step 1 in PR #14626)
2. Use array().abs() + kEpsilon instead of for{for{abs() + kEpsilon}}. (steps 1+2 in PR #14628)
3. Use matrix-vector multiplications in Ax,Ay,Az directions. (this PR #14629 is steps 1+2+3)

This PR have steps 1+2+3 above.  It didn't help improving the running time compared to 1+2 probably because we are not exploiting AVX in step 3 yet.
```
drake (ObbOptimize3AxAyAz)$ geometry/benchmarking/obb_benchmark/conduct_experiment ~/project/ObbBench/2021-02-06_ObbOptimize3AxAyAz
$ tail -n12 ~/project/ObbBench/2021-02-06_ObbOptimize3AxAyAz/summary.txt
-------------------------------------------------------------------------------------
Benchmark                                           Time             CPU   Iterations
-------------------------------------------------------------------------------------
BM_ObbHasOverlap_31_Cases_mean                   2985 ns         2985 ns            9
BM_ObbHasOverlap_31_Cases_median                 2985 ns         2984 ns            9
BM_ObbHasOverlap_31_Cases_stddev                 1.14 ns         1.14 ns            9
BM_ObbHasOverlap_16_OverlapCases_mean            1723 ns         1723 ns            9
BM_ObbHasOverlap_16_OverlapCases_median          1723 ns         1723 ns            9
BM_ObbHasOverlap_16_OverlapCases_stddev         0.570 ns        0.505 ns            9
BM_ObbHasOverlap_15_NonOverlapCases_mean         1391 ns         1391 ns            9
BM_ObbHasOverlap_15_NonOverlapCases_median       1392 ns         1392 ns            9
BM_ObbHasOverlap_15_NonOverlapCases_stddev      0.488 ns        0.473 ns            9
```

<!-- Reviewable:start -->
---
This change is [<img src="https://reviewable.io/review_button.svg" height="34" align="absmiddle" alt="Reviewable"/>](https://reviewable.io/reviews/robotlocomotion/drake/14629)
<!-- Reviewable:end -->